# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: lycmap.ai Study AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quiz/generate:
    post:
      operationId: generateQuiz
      tags: [quiz]
      summary: Generate a quiz based on board, class, subject, chapter, level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizRequest"
      responses:
        "200":
          description: Generated quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
...
  /quiz/sessions/{id}:
    get:
      operationId: getQuizSession
      tags: [quiz]
      summary: Get a quiz session with all questions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /quiz/sessions/{id}/complete:
    post:
      operationId: completeQuizSession
      tags: [quiz]
      summary: Mark a quiz session as completed with score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizCompletion"
      responses:
        "200":
          description: Completed session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizSession"
...
  /quiz/formula-sheet:
    post:
      operationId: getFormulaSheet
      tags: [quiz]
      summary: Get formula sheet for a subject/chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterRequest"
      responses:
        "200":
          description: Formula sheet data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormulaSheet"
  /quiz/short-notes:
    post:
      operationId: getShortNotes
      tags: [quiz]
      summary: Get short notes for a subject/chapter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChapterRequest"
      responses:
        "200":
          description: Short notes data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShortNotes"
  /quiz/upload:
    post:
      operationId: uploadDocument
      tags: [quiz]
      summary: Upload a document/image for AI analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "200":
          description: Upload analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnalysis"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    QuizRequest:
      type: object
      required: [board, className, subject, chapter, level]
      properties:
        board:
          type: string
          description: "CBSE, ICSE, SCERT, or Mixed"
        className:
          type: string
          description: "Class 6 through Class 12"
        subject:
          type: string
        chapter:
          type: string
        level:
          type: string
          description: "Easy, Medium, Hard, or Mixed"
        questionCount:
          type: integer
          default: 20
    ChapterRequest:
      type: object
      required: [board, className, subject, chapter]
      properties:
        board:
          type: string
        className:
          type: string
        subject:
          type: string
        chapter:
          type: string
    QuizQuestion:
      type: object
      required: [id, questionText, options, correctAnswer, explanation, type]
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        type:
          type: string
          description: "previous_year or ai_generated"
        year:
          type: ["string", "null"]
        topic:
          type: string
        difficulty:
          type: string
    QuizSession:
      type: object
      required: [id, board, className, subject, chapter, level, status, createdAt, questions]
      properties:
        id:
          type: integer
        board:
          type: string
        className:
          type: string
        subject:
          type: string
        chapter:
          type: string
        level:
          type: string
        status:
          type: string
          description: "in_progress or completed"
        score:
          type: ["integer", "null"]
        totalQuestions:
          type: integer
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
...
    QuizCompletion:
      type: object
      required: [score]
      properties:
        score:
          type: integer
...
    FormulaFormula:
      type: object
      required: [name, formula, description, unit]
      properties:
        name:
          type: string
        formula:
          type: string
        description:
          type: string
        unit:
          type: string
        derivation:
          type: ["string", "null"]
    FormulaSection:
      type: object
      required: [sectionTitle, formulas]
      properties:
        sectionTitle:
          type: string
        formulas:
          type: array
          items:
            $ref: "#/components/schemas/FormulaFormula"
    FormulaSheet:
      type: object
      required: [subject, chapter, sections, previousYearTopics]
      properties:
        subject:
          type: string
        chapter:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/FormulaSection"
        previousYearTopics:
          type: array
          items:
            type: string
    ShortNotesItem:
      type: object
      required: [heading, content, keyPoints]
      properties:
        heading:
          type: string
        content:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        equations:
          type: array
          items:
            type: string
        previousYearRelevance:
          type: ["string", "null"]
    ShortNotes:
      type: object
      required: [subject, chapter, items, previousYearGraph]
      properties:
        subject:
          type: string
        chapter:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/ShortNotesItem"
        previousYearGraph:
          type: array
          items:
            $ref: "#/components/schemas/PreviousYearDataPoint"
    PreviousYearDataPoint:
      type: object
      required: [year, questionsCount, topics]
      properties:
        year:
          type: string
        questionsCount:
          type: integer
        topics:
          type: array
          items:
            type: string
    DocumentUpload:
      type: object
      required: [fileName, fileData, mimeType]
      properties:
        fileName:
          type: string
        fileData:
          type: string
          description: Base64-encoded file content
        mimeType:
          type: string
    DocumentAnalysis:
      type: object
      required: [extractedText, detectedSubject, detectedChapter, suggestedQuestions]
      properties:
        extractedText:
          type: string
        detectedSubject:
          type: ["string", "null"]
        detectedChapter:
          type: ["string", "null"]
        suggestedQuestions:
          type: array
          items:
            type: string
...