# lib/api-zod/src/generated/types/documentUpload.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentUpload {
  fileName: string;
  /** Base64-encoded file content */
  fileData: string;
  mimeType: string;
}