# lib/api-zod/src/generated/types/formulaSection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */
import type { FormulaFormula } from './formulaFormula';

export interface FormulaSection {
  sectionTitle: string;
  formulas: FormulaFormula[];
}