# lib/api-zod/src/generated/types/formulaSheet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */
import type { FormulaSection } from './formulaSection';

export interface FormulaSheet {
  subject: string;
  chapter: string;
  sections: FormulaSection[];
  previousYearTopics: string[];
}