# lib/api-zod/src/generated/types/previousYearDataPoint.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface PreviousYearDataPoint {
  year: string;
  questionsCount: number;
  topics: string[];
}