# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  id: number;
  sessionId?: number;
  questionText: string;
  options: string[];
  correctAnswer: string;
  explanation: string;
  /** previous_year or ai_generated */
  type: string;
  /** @nullable */
  year?: string | null;
  topic?: string;
  difficulty?: string;
}