# lib/api-zod/src/generated/types/quizRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface QuizRequest {
  /** CBSE, ICSE, SCERT, or Mixed */
  board: string;
  /** Class 6 through Class 12 */
  className: string;
  subject: string;
  chapter: string;
  /** Easy, Medium, Hard, or Mixed */
  level: string;
  questionCount?: number;
}