# lib/api-zod/src/generated/types/shortNotes.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */
import type { PreviousYearDataPoint } from './previousYearDataPoint';
import type { ShortNotesItem } from './shortNotesItem';

export interface ShortNotes {
  subject: string;
  chapter: string;
  items: ShortNotesItem[];
  previousYearGraph: PreviousYearDataPoint[];
}