# lib/api-zod/src/generated/types/shortNotesItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * lycmap.ai Study AI API
 * OpenAPI spec version: 0.1.0
 */

export interface ShortNotesItem {
  heading: string;
  content: string;
  keyPoints: string[];
  equations?: string[];
  /** @nullable */
  previousYearRelevance?: string | null;
}